refactor tests and organize folders
