+++ server/api/model/person.js
@@ -11,17 +11,21 @@ const Exception = require('./exception');
 const seconds = 1000;
 
 class Person extends BaseModel {
-	static get rules () {
+	get rules () {
 		return {
 			email: ['required', 'email'],
 			password: ['required', 'string'],
-			status: ['required', 'integer'],
+			status: 'integer',
 			admin: ['integer']
 		};
 	}
 
 	get hidden () {
 		return ['password', 'admin', 'status'];
+	}
+
+	get hasTimestamps() {
+		return true;
 	}
 
 	get tableName () {
