+++ server/package.json
@@ -64,7 +64,7 @@
         "build": "./node_modules/react-scripts/bin/react-scripts.js build",
         "serve": "node bin/serve",
         "lint": "./node_modules/.bin/eslint . --fix",
-        "test": "NODE_ENV=test ./node_modules/.bin/mocha ./test/**/*.js --exit --recursive --require ./test/setup.js"
+        "test": "NODE_ENV=test ./node_modules/.bin/mocha ./tests/integration/* ./tests/unit/* --exit --recursive --require ./tests/setup.js"
     },
     "devDependencies": {
         "@typescript-eslint/eslint-plugin": "^3.9.1",
