+++ server/tests/tables_structs/alert_struct.js
