+++ server/tests/tables_structs/archive_notification_struct.js
