+++ server/tests/tables_structs/notification_struct.js
